Fixed up OpenAL as it wasn't streaming properly. Fixed up bits of coding for FMOD, Mpg123 and Sound.

+++ src/sound/mpg123_decoder.h
@@ -7,15 +7,17 @@
 #ifdef HAVE_MPG123
 
 #ifdef _MSC_VER
-typedef int ssize_t;
+#include <stddef.h>
+typedef ptrdiff_t ssize_t;
 #endif
+
 #include "mpg123.h"
 
 struct MPG123Decoder : public SoundDecoder
 {
-    virtual void getInfo(int *samplerate, ChannelConfig *chans, SampleType *type);
+    virtual void getInfo(int* samplerate, ChannelConfig* chans, SampleType* type);
 
-    virtual size_t read(char *buffer, size_t bytes);
+    virtual size_t read(char* buffer, size_t bytes);
     virtual bool seek(size_t ms_offset);
     virtual size_t getSampleOffset();
     virtual size_t getSampleLength();
@@ -24,20 +26,20 @@ struct MPG123Decoder : public SoundDecoder
     virtual ~MPG123Decoder();
 
 protected:
-    virtual bool open(FileReader *reader);
+    virtual bool open(FileReader* reader);
 
 private:
-    mpg123_handle *MPG123;
+    mpg123_handle* MPG123;
     bool Done;
 
-    FileReader *Reader;
+    FileReader* Reader;
     int StartOffset;
-    static off_t file_lseek(void *handle, off_t offset, int whence);
-    static ssize_t file_read(void *handle, void *buffer, size_t bytes);
+    static off_t file_lseek(void* handle, off_t offset, int whence);
+    static ssize_t file_read(void* handle, void* buffer, size_t bytes);
 
     // Make non-copyable
-    MPG123Decoder(const MPG123Decoder &rhs);
-    MPG123Decoder& operator=(const MPG123Decoder &rhs);
+    MPG123Decoder(const MPG123Decoder& rhs);
+    MPG123Decoder& operator=(const MPG123Decoder& rhs);
 };
 
 #endif
